web-nfc/nfc_push_ignoreRead_false-manual.https.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Web NFC Test: push message with false ignoreRead</title>
<link rel="author" title="Intel" href="http://www.intel.com"/>
<link rel="help" href="https://w3c.github.io/web-nfc/"/>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/nfc_help.js"></script>
<meta name="flags" content="interact">
<meta name="timeout" content="long">

<p>Tap an NFC tag to the test device with NFC support.</p>

<p>Note: All the actions need to be done in 60 seconds, otherwise it will get TIMEOUT.</p>

<div id="log"></div>

<script>

"use strict";

setup({ explicit_timeout: true });

promise_test(t => {
  return navigator.nfc.push(test_text_data)
    .then(() => {
      return new Promise(resolve => {
        navigator.nfc.watch(message => resolve(message));
        navigator.nfc.push("This message will not be watched", {ignoreRead: false}).then(()=>console.log("push twice"));
      }).then((message) => {
            for (let record of message.records) {
              console.log("data::::: ", record.data)
              assert_equals(record.data, test_text_data);
            }
          });
      });
}, "Test that nfc.push succeeds when ignoreRead is set to false.");

</script>
